ENG-7251 - Refactor RN Events folder to deduplicate code
