+++ NeuroID/src/reactNative/java/com/neuroid/tracker/events/NIDRegisterEventsHelper.kt
@@ -1,63 +1,10 @@
 package com.neuroid.tracker.events
 
-import android.app.Activity
 import android.os.Looper
-import android.view.View
-import android.view.ViewGroup
-import com.neuroid.tracker.callbacks.*
-import com.neuroid.tracker.storage.NIDDataStoreManager
-import com.neuroid.tracker.utils.NIDLogWrapper
-import java.util.*
 
-fun registerWindowListeners(activity: Activity) {
-    val viewMainContainer = activity.window.decorView.findViewById<View>(
-        android.R.id.content
-    )
-
-    val callBack = activity.window.callback
-
-    if (callBack !is NIDGlobalEventCallback) {
-        val nidGlobalEventCallback = NIDGlobalEventCallback(
-            callBack,
-            NIDTouchEventManager(viewMainContainer as ViewGroup),
-            viewMainContainer
-        )
-        viewMainContainer.viewTreeObserver.addOnGlobalFocusChangeListener(nidGlobalEventCallback)
-        viewMainContainer.viewTreeObserver.addOnGlobalLayoutListener(
-            nidGlobalEventCallback
-        )
-
-        activity.window.callback = nidGlobalEventCallback
-    }
-}
-
-fun registerTargetFromScreen(
-    activity: Activity,
-    logger: NIDLogWrapper,
-    storeManager: NIDDataStoreManager,
-    registerTarget: Boolean,
-    registerListeners: Boolean,
-    activityOrFragment: String = "",
-    parent: String = "",
-) {
-    val viewMainContainer = activity.window.decorView.findViewById<View>(
-        android.R.id.content
-    ) as ViewGroup
-
-    val hashCodeAct = activity.hashCode()
-    val guid = UUID.nameUUIDFromBytes(hashCodeAct.toString().toByteArray()).toString()
-
+// run the function after a 300ms delay in react native
+internal fun handleIdentifyAllViews(r:Runnable){
     android.os.Handler(Looper.getMainLooper()).postDelayed({
-        identifyAllViews(
-            viewMainContainer,
-            guid,
-            logger,
-            storeManager,
-            registerTarget,
-            registerListeners,
-            activityOrFragment,
-            parent
-        )
+        r.run()
     }, 300)
 }
-
